Update shubox to use new api (#913)
What changed?
=============

Constable was slow to load, seemingly in part because of shubox. So we
move the loading of shubox to be async in a055a97231bc4b1a13e7b50c739ad0c1809082db.

While researching that, we noticed we have an old way of using shubox.
This commit updates shubox to the new way:

1. Install it via npm (no longer need to hit a specific url
asynchronously)

2. Update the `s3urlTemplate` to `successTemplate` init option. The
`s3urlTemplate` is deprecated. See
https://github.com/shuboxio/shubox.js/blob/master/README.md#successtemplate

3. Update configuration files, sample.env, etc. to refer to `SHUBOX_KEY`
and `shubox_key` instead of `shubox_script_url`. Shubox now has a unique
key for each environment/app, so we need to use the correct ones. Note
that there is a sandbox mode for development: so if desired,
developers/designers can put the sandbox key in their .env files and
upload images locally. All keys are available in shubox's dashboard.

There are some improvements we could make. In particular, we could add
an `uploadingTemplate` to shubox to notify the user that we're in the
process of uploading the image. But we found some difficulty with that
and we leave that work for the future.