Add "Uploading ..." placeholder when uploading images (#927)
Closes https://github.com/thoughtbot/constable/issues/915

What changed?
============

We add an "Uploading ..." placeholder (it actually includes the
filename) when uploading images into a constable post. Currently, when
uploading an image, it is hard to know if the dropzone captured the
image or not because the user gets no feedback about the upload being in
progress. Shubox has an `uploadingTemplate` option but there was a bug
that was fixed in version `0.5.0`. So we update our Shubox version and
make use of the `uploadingTemplate` option.