Add Phoenix LiveView
What changed?
============

We add Phoenix LiveView, following instructions in
https://hexdocs.pm/phoenix_live_view/installation.html

We had to break apart a root layout, which caused to change some of the
work done in https://github.com/thoughtbot/constable/pull/745 to reduce
the number of layouts.

We reintroduce the "login.html" layout, but since our layouts now share
a root layout, "app.html", "live.html", and "login.html" are very
minimal and share most of the code via the "root.html" layout.

We also put the "/recipients_preview" route under the `:api` umbrella
since it renders json, and we now have a root layout in the `:browser`
pipeline. Otherwise, it requires a `root.json` layout to exist and break
our tests.

Co-authored-by: Mike Burns <mburns@thoughtbot.com>
Co-authored-by: Brian Tenggren <tenggren@thoughtbot.com>