Create announcement from AnnouncementLive
What changed?
=============

Having moved the logic previously found in `announcements/new` to the
`AnnouncementLive` live view, we now move the actual creation into
`AnnouncementLive`. Thus, we can get rid of the `create` action in the
`AnnouncementController` (as well as the `new` action which we could've
already done).

To create an announcement, we need to know who the current user is. So
we also update the `FetchCurrentUser` plug to set the `current_user_id`
in session in order to have access to it from our live view.

In order to properly show an error message when no interests are
selected, we update the `new.html` to wrap the selectize portion and the
error tag within one more div. This allows us to specify the
`phx-update="ignore"` attribute on the label and input (since both seem
to need it), but it let's live view update the error tag when the
changeset returns errors for the empty interests.