Use LiveView in interests index page
What changed?
============

We move the interests' index page to `InterestLive`. The current
experience is a bit clunky: subscribing to an interest far down the page
reloads the page causing it to go back up to the top. With LiveView we
can handle that more smoothly.

In addition to moving the create/delete of user-interests to
`InterestLive`, we also set the `user_id` in session (to get it from the
live view). Doing so also revealed a little but in `FetchCurrentUser`
plug that assumed we weren't setting a `current_user` assign with an
inactive user, but it was actually setting a `nil` user. We now
explicitly handle that.

Finally, LiveView does not play nicely with Turbolinks, since Tubolinks
doesn't do a full page load and the `InterestLive` is mounted on the
router. So, we do not use Turbolinks to go to the interest's index page
since that prevents the correct loading of live view.

Co-authored-by: Mike Burns <mburns@thoughtbot.com>