Render announcement through AnnouncementShowLive
What changed?
===========

We update the announcement page to be a live view rather than a static
page. That will allow us to move commenting completely inside the
live view along with the subscribe/unsubscribe from the announcement
(which currently has a bug).

To do that, we render the `AnnouncementShowLive` from
`AnnouncementController#show`.

We could have rendered the live view page directly from the router, but
we render it from the controller because:

- the controller already has the current user, so we can simply pass it
  as session data, and
- the controller does the deslugifying of the `id`, so we can keep that
  logic centralized in the `AnnouncementController`.

Creating comments
---------------

A user can now create comments directly via the `AnnouncementShowLive`,
meaning we no longer have to make an AJAX request to render the comment,
and we no longer have to receive the HTML over the wire via the
"live-html" channel.

Receiving new comments in real-time
----------------------------------

Users would already receive new comments in real-time. We keep that
behavior but we change the implementation. Prior to this commit, a
browser would subscribe to a "live-html" channel and receive updates.
That's how the server would send the new comment HTML, and the JS would
append the comment to the list of comments.

Now that we have the `AnnouncementShowLive`, we can simplify that by
letting the live view push an updated list of comments whenever there is
a new comment. To do that, we subscribe the `AnnouncementShowLive` to
the announcement's topic via Phoenix PubSub. So, now when an
announcement is created, rather than publishing updates via the
Endpoint, we can use Phoenix PubSub to broadcast the update to all
subscribers for that announcement.

Because the "live-html" channel (and in fact the "update" channel) were
only used for sending the new comments over the wire, and since we no
longer need them, we can now remove them completely.

Comment-form.js updates
-----------------------

Since we no longer need to submit the comment via AJAX, the
`comment-form.js` file is substantially diminished. We still do some
things via JavaScript, so we do not remove the file. But the form
submission no longer happens through that.

What we still do via JS:

- We still do the comment markdown processing
- We still do the toggling between compose & preview for the comment
- We still keep a draft of the comment in localStorage
- We still do image uploading via JS
- We still do the user mentions via JS

Because we're using JS with LiveView, we move things as much as possible
into the preferred way to interop between LiveView and JS -- namely, we
use Phoenix LiveView hooks.

For example, we move the `ImageUploader` into a Phoenix LiveView Hook. A
benefit of that is that we can make it agnostic of the element's DOM id.
We only have to set the `ImageUploader` hook on any element (via a
`phx-hook`). So even though we only use it for comment forms right now,
we can use it in the future with announcements.

The one thing we do not do via a JS hook is the user mentions, and that
only because we need the users list on page load.

Turbolinks note
----------------

A known issue between turbolinks and LiveView is that if we navigate to
a page via turbolinks (and thus do not load the page completely),
LiveView isn't mounted properly. Thus, we disable turbolinks (via
`data-turbolinks-false` attribute) when navigating to the announcement
show page.