Do not use ConstableWeb, :view in ConnCase (#980)
What changed?
============

We remove the `use ConstableWeb, :view` from our `ConnCase` test helper.

Using `ConstableWeb, :view` brings in a lot of helpers, but it's meant
to be used from Phoenix views, not in conn helpers. Putting it in
`ConnCase` effectively turns every test that uses `ConnCase` into a
Phoenix view.

It seems we haven't had issues with that so far, but in future work (in
a separate branch), I ran into issues because Phoenix views import a
`render` function.

So, instead of `use`ing `ConstableWeb, :view`, I found the test that
needed the helpers (EmailTest) and imported the `link` and
`safe_to_string` helpers directly.