Use full URLs for profile and image in User factory
What changed?
-----------

We use full URLs for profile URLs and image URLs in the user factory.

Why?
----

Not doing so makes it so that the paths in HTML templates (in tests) are
relative paths (e.g. /announcemets/someurlimage.com). That results in
odd tests failures at times when websockets are involved.