Fix View Constable link in emails
What changed?
============

We fix a "View Constable" in new announcement emails that went
completely vertical when viewed in an iphone.

We limit the width of the interests to 50%, and we use a deprecated (but
still used) table attribute `valign` to align the "View on Constable" to
the top of those interests.

Co-authored-by: Eric Bailey <eric.w.bailey@gmail.com>