Do not display markdown & preview side by side
What changed?
============

When composing a new announcement, we show the user both the markdown
they're typing as well as the preview (in realtime). Though the
experience looks cool, it makes it more difficult to compose
announcements. That is particularly true if you aren't using the full
screen of a large monitor.

On a mobile screen, the markdown and preview screens are hidden behind
two tabs that can be clicked to toggle the state. This seems like a
nicer compose view (in my opinion).

But there are many screen sizes in between the full monitor (when
markdown and preview look okay) and a small mobile screen (where
markdown and preview use tabs).

This commit changes the announcement new and edit pages to always use
the tabbed view. That allows the author to focus on the composing the
announcement in the center of the page, but still shows the preview on
demand when the "preview" tab is clicked.

Implementation notes
----------------------

As part of this work, we move the code found in
`announcement-form-mobile.js` into the `announcement-form.js`, since we
no longer make a distinction between the regular and mobile forms. The
regular form should use the tabs just like the mobile form did before.